Add combos for Callum style mod activation
This attempts to solve the problem of mods not firing if you cord the mod
activation with the mod itself. With these combos enabled, cording the mod and
the activation key / clutch will still activate the mod. Other mods can be
pressed after the initial combo as long as the combo itself remains held.

Remaining issues:
  - Releasing either of the combo keys deactivates the mods layer
  - Multiple macros don't fire at the same time

I'm not sure how easy it will be to solve either of these issues. The first may
be a limitation of zmk. The second can probably be solved by adding one combo
for each combination.